apps/speech/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
faster-whisper==1.0.3
ollama==0.4.5
requests==2.32.3